(test) Move reducer tests to own file
