+++ test/client/home.spec.js
@@ -6,50 +6,12 @@ import { renderIntoDocument, scryRenderedDOMComponentsWithTag, Simulate } from '
 import { expect } from 'chai';
 import axios from 'axios';
 
-import reducer from '../../client/reducers';
 import { Home } from '../../client/components/Home';
 
 
-var chai = require('chai');
-
-describe('Home Auth Reducers', () => {
-
-  it('handles FETCH_PULSE', () => {
-    const initialState = { pulse: 0, logState: false };
-    const action = {
-      type: 'FETCH_PULSE',
-      payload: 1
-    }
-    const nextState = reducer(initialState, action);
-    expect(nextState.pulse).to.equal(1);
-    expect(nextState.logState).to.equal(false);
-  })
-
-  it('handles FETCH_LOGSTATE', () => {
-    const initialState = { pulse: 0, logState: false };
-    const action = {
-      type:'FETCH_LOGSTATE',
-      payload: { data: true }
-    }
-    const nextState = reducer(initialState, action);
-    expect(nextState.pulse).to.equal(0);
-    expect(nextState.logState).to.equal(true);
-  })
-
-  it('handles undefined state', () => {
-    const action = {
-      type: 'FETCH_LOGSTATE',
-      payload: { data: true }
-    }
-    const nextState = reducer(undefined, action);
-    expect(nextState.pulse).to.equal(0);
-    expect(nextState.logState).to.equal(true);
-  })
-})
-
 describe('Home Auth Component', () => {
 
-  it('renders a button and a link when logged out', () => {
+  xit('renders a button and a link when logged out', () => {
     let logState = false;
     const fetchLogState = () => logState = false;
     const component = renderIntoDocument(
@@ -67,7 +29,7 @@ describe('Home Auth Component', () => {
 
   })
 
-  it('renders 2 buttons when logged in',() => {
+  xit('renders 2 buttons when logged in',() => {
     let logState = true;
     const fetchLogState = () => logState = true;
     const component = renderIntoDocument(
